third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-in-detached-frame-expected.txt

Tests navigator.requestMIDIAccess fails in a detached frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS errorName is "AbortError"
PASS requestMIDIAccess() fail as expected.
PASS successfullyParsed is true

TEST COMPLETE
